Assignment 3: Fix warnings

+++ 2013 Winter/Matchismo/Matchismo Tests/SetCardTests.m
@@ -30,7 +30,7 @@
 
 #pragma mark - Helpers
 
-- (SetCard *)createCardWithNumber:(int)number
+- (SetCard *)createCardWithNumber:(NSUInteger)number
                           shading:(CardShadingType)shading
                             color:(CardColorType)color
                            symbol:(NSString *)symbol
