+++ 2013 Winter/Matchismo/Matchismo/Components/PlayingDeck.m
@@ -16,7 +16,7 @@
     self = [super init];
     if (self) {
         for (NSString *suit in [PlayingCard validSuits]) {
-            for (int rank = 1; rank <= [PlayingCard maxRank]; rank++) {
+            for (NSUInteger rank = 1; rank <= [PlayingCard maxRank]; rank++) {
                 PlayingCard *card = [[PlayingCard alloc] init];
                 card.suit = suit;
                 card.rank = rank;
